Refactored AWS EC2 utility functions and relocated tests.
**Added:**

- `network.go` - Introduced new file for networking-related functions.
- `GetSubnetID` function - Retrieves the ID of the subnet by name.
- `GetVPCID` function - Fetches the ID of the VPC by name.
- `IsSubnetPubliclyRoutable` function - Checks if a subnet is publicly routable.
- `ListSecurityGroupsForVpc` function - Lists all security groups for a VPC ID.
- `network_test.go` - Added tests for newly relocated functions.

**Changed:**

- Relocated `GetSubnetID`, `GetVPCID`, `IsSubnetPubliclyRoutable`, and
  `ListSecurityGroupsForVpc` from `ec2.go` to `network.go`.
- Refactored associated tests to match the new file structure.
- Updated the import paths and references in tests to align with the new file
  structure.

**Removed:**

- Redundant code from `ec2.go` and `ec2_test.go` after the relocation of
  networking-related functions and tests.
