+++ ec2/ec2_test.go
@@ -372,146 +372,6 @@ func TestListSecurityGroupsForSubnet(t *testing.T) {
 	}
 }
 
-func TestIsSubnetPubliclyRoutable(t *testing.T) {
-	c := ec2utils.NewConnection()
-	routableSubnetID, err := c.GetSubnetID("test-subnet-2")
-	if err != nil {
-		t.Fatalf("failed to get VPC ID: %v", err)
-	}
-
-	tests := []struct {
-		name      string
-		subnetID  string
-		expectErr bool
-	}{
-		{
-			name:      "Valid Subnet ID",
-			subnetID:  routableSubnetID,
-			expectErr: false,
-		},
-		{
-			name:      "Invalid Subnet ID",
-			subnetID:  "subnet-invalid",
-			expectErr: true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			c := ec2utils.NewConnection()
-			_, gotError := c.IsSubnetPubliclyRoutable(tc.subnetID)
-
-			if tc.expectErr {
-				assert.Error(t, gotError)
-			} else {
-				assert.NoError(t, gotError)
-			}
-		})
-	}
-}
-
-func TestListSecurityGroupsForVpc(t *testing.T) {
-	c := ec2utils.NewConnection()
-	validVPCID, err := c.GetVPCID("test-vpc")
-	if err != nil {
-		t.Fatalf("failed to get VPC ID: %v", err)
-	}
-
-	tests := []struct {
-		name      string
-		vpcID     string
-		expectErr bool
-	}{
-		{
-			name:      "Valid VPC ID",
-			vpcID:     validVPCID,
-			expectErr: false,
-		},
-		{
-			name:      "Invalid VPC ID",
-			vpcID:     "vpc-invalid",
-			expectErr: true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			c := ec2utils.NewConnection()
-			_, gotError := c.ListSecurityGroupsForVpc(tc.vpcID)
-
-			if tc.expectErr {
-				assert.Error(t, gotError)
-			} else {
-				assert.NoError(t, gotError)
-			}
-		})
-	}
-}
-
-func TestGetSubnetID(t *testing.T) {
-	tests := []struct {
-		name       string
-		subnetName string
-		expectErr  bool
-	}{
-		{
-			name:       "Valid Subnet Name",
-			subnetName: "test-subnet-2",
-			expectErr:  false,
-		},
-		{
-			name:       "Invalid Subnet Name",
-			subnetName: "InvalidSubnet",
-			expectErr:  true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			c := ec2utils.NewConnection()
-			_, gotError := c.GetSubnetID(tc.subnetName)
-
-			if tc.expectErr {
-				assert.Error(t, gotError)
-			} else {
-				assert.NoError(t, gotError)
-			}
-		})
-	}
-}
-
-func TestGetVPCID(t *testing.T) {
-	tests := []struct {
-		name      string
-		vpcName   string
-		expectErr bool
-	}{
-		{
-			name:      "Valid VPC Name",
-			vpcName:   "test-vpc",
-			expectErr: false,
-		},
-		{
-			name:      "Invalid VPC Name",
-			vpcName:   "InvalidVPC",
-			expectErr: true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			c := ec2utils.NewConnection()
-			_, gotError := c.GetVPCID(tc.vpcName)
-
-			if tc.expectErr {
-				assert.Error(t, gotError)
-			} else {
-				assert.NoError(t, gotError)
-			}
-		})
-	}
-}
-
 func TestCreateSecurityGroup(t *testing.T) {
 	c := ec2utils.NewConnection()
 	vpcID, err := c.GetVPCID("default")
